修正：レビュー対応
